backend/requirements/base.txt
# Python
pre-commit==2.15.0 # https://pre-commit.com/
requests==2.28.0 # https://pypi.org/project/requests/

# Django
Django==3.2.16
django-smart-selects==1.5.9

django-extensions==3.1.5 # https://django-extensions.readthedocs.io/en/latest/installation_instructions.html
django-import-export==2.8.0 # https://django-import-export.readthedocs.io/en/latest/installation.html
django-debug-toolbar==3.5.0 # https://django-debug-toolbar.readthedocs.io/en/latest/installation.html
django-filter==22.1
pytest-django # https://pytest-django.readthedocs.io/en/latest/
pytest-cov
django-phonenumber-field==6.1.0
django-ckeditor==6.5.1
django-recaptcha==3.0.0

# Postgres
psycopg2-binary==2.9.2 # https://pypi.org/project/psycopg2-binary/

Pillow==9.2.0 #pip install Pillow
django-crispy-forms==1.14.0 #pip install django-crispy-forms


# extra
six